Title: 합과 차의 나눗셈, Time: 132ms, Memory: 10MB, Status: Passed - Codetree

+++ 251020/합과 차의 나눗셈/divide-of-sum-and-sub.java
@@ -6,7 +6,7 @@ public static void main(String[] args) {
         int a = sc.nextInt();
         int b = sc.nextInt();
 
-        System.out.printf("%.2f",(a+b)/(a-b));
+        System.out.printf("%.2f",(double)(a+b)/(a-b));
 
 
     }
